Improve get_ai_move() function
The function first checks all rows combinations (for both X and O)
and save potential best move into a dictionary 'best_move' that
stores information about coordinates of the best move and the
highest number of marks in a row. In the next step the function
returns the coordinates of the best move.